# src/swagger/documentation.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is Confi API Documentation",
    "version": "1.0.0",
    "title": "Swagger Confi",
    "contact": {
      "email": "aloncar.mail@gmail.com"
    }
  },
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
    "/app/user": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Add a new user",
        "description": "",
        "operationId": "addUser",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "email",
            "description": "User email which is necessary for login.",
            "required": true
          },
          {
            "in": "body",
            "name": "password",
            "description": "User password which is necessary for login.",
            "required": true
          },
          {
            "in": "body",
            "name": "isAdmin",
            "description": "User 'is admin' flag.",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "300": {
            "description": "Required params not found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Fetch an existing user",
        "description": "",
        "operationId": "getUser",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "session",
            "name": "userId",
            "description": "Logged user ID.",
            "required": true
          }
        ],
        "responses": {
          "300": {
            "description": "Required params not found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "400": {
            "description": "User not found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        },
        "security": [
          {
            "auth": [
              "token"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "user"
        ],
        "summary": "Delete an existing user",
        "description": "",
        "operationId": "deleteUser",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "session",
            "name": "userId",
            "description": "Logged user ID.",
            "required": true
          }
        ],
        "responses": {
          "300": {
            "description": "Required params not found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "400": {
            "description": "User not found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "200": {
            "description": "User deleted",
            "schema": {
              "type": "object",
              "properties": {
                "ok": {
                  "type": "boolean"
                },
                "isDeleted": {
                  "type": "boolean"
                },
                "user": {
                  "$ref": "#/definitions/User"
                }
              }
            }
          }
        },
        "security": [
          {
            "auth": [
              "token"
            ]
          }
        ]
      }
    },
...
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "isAdmin": {
          "type": "boolean"
        }
      }
    },
...
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "ok": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        }
      }
    }
  },
...
}